First time deployment to heroku

+++ admin/search.php
@@ -1,43 +1,8 @@
 <form action="welcome.php">
-    <input class="button dashboardbutton" type="submit" value="Go to Dashboard" />
+    <input class="button dashboardbutton" type="submit" value="Go Back" />
 </form>
 <link rel="stylesheet" href="../css/style.css">
-<style>
-  .dashboardbutton {
-  background-color: white;
-  color: black;
-  border: 2px solid #e7e7e7;
-  padding: 10px 24px;
-  transition-duration: 0.4s;
-}
-.dashboardbutton:hover{
-  background-color: #e7e7e7;
-}
-
-#studenttable {
-  font-family: Arial, Helvetica, sans-serif;
-  border-collapse: collapse;
-  width: 100%;
-}
-
-#studenttable td, #studenttable th {
-  border: 1px solid #ddd;
-  padding: 8px;
-}
-
-#studenttable tr:nth-child(even){background-color: #f2f2f2;}
-
-#studenttable tr:hover {background-color: #ddd;}
-
-#studenttable th {
-  padding-top: 12px;
-  padding-bottom: 12px;
-  text-align: left;
-  background-color: #04AA6D;
-  color: white;
-}
-
-</style>
+<link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.4.1/css/bootstrap.min.css">
 
 <?php
 require_once('connect.php');
@@ -45,35 +10,41 @@ if (isset($_POST['search'])){
     $valueToSearch = $_POST['valueToSearch'];
     
     $query= "SELECT * FROM articles_table
-    WHERE title LIKE '%$valueToSearch%' OR author LIKE '%$valueToSearch%' OR subj LIKE '%$valueToSearch%' OR publisher LIKE '%$valueToSearch%'";
+    WHERE title LIKE '%$valueToSearch%' OR abstract LIKE '%$valueToSearch%' OR subj LIKE '%$valueToSearch%'";
     $result = @mysqli_query($dbc, $query);
    
     if (mysqli_num_rows($result) > 0) {
         
         echo "<table border='1' id='studenttable'>
         <tr>
-               <th>ID</th>
-               <th>Title</th>
-               <th>Author</th>
-               <th>Subject</th>
-               <th>Publisher</th>
-               <th>Publication Date</th>
-               <th>Call Number</th>
-               <th>Accession Number</th>
+        <th>ID</th>
+        <th>Title</th>
+        <th>Subject</th>
+        <th>Publisher</th>
+        <th>Publication Date</th>
+        <th>Publication Place</th>
+        <th>Volume</th>
+        <th>Number</th>
+        <th>Series</th>
+        <th>Pages</th>
+        <th>Abstract</th>
                <th>Actions </th>
                </tr>";
         
         while($row = @mysqli_fetch_array($result)) { ?>
             <tr>
-                  <td> <?php echo $row['id'] ?> </td>
+            <td> <?php echo $row['id'] ?> </td>
                   <td> <?php echo $row['title'] ?> </td>
-                  <td> <?php echo $row['author'] ?> </td>
                   <td> <?php echo $row['subj'] ?> </td>
                   <td> <?php echo $row['publisher'] ?> </td>
-                  <td> <?php echo $row['publication_date'] ?> </td>
-                  <td> <?php echo $row['call_number'] ?> </td>
-                  <td> <?php echo $row['accession_number'] ?> </td>
-                  <td> <a class="btn" href="update.php?id=<?php echo $row['id']?>">Update</a> <a class="btn" href="delete.php?id=<?php echo $row['id']?>">Delete</a></td>
+                  <td> <?php echo $row['publicationdate'] ?> </td>
+                  <td> <?php echo $row['publicationplace'] ?> </td>
+                  <td> <?php echo $row['volume'] ?> </td>
+                  <td> <?php echo $row['num'] ?> </td>
+                  <td> <?php echo $row['series'] ?> </td>
+                  <td> <?php echo $row['pages'] ?> </td>
+                  <td> <?php echo $row['abstract'] ?> </td>
+                  <td> <a class="btn btn-primary" href="update.php?id=<?php echo $row['id']?>">Update</a> <a class="btn btn-danger" href="delete.php?id=<?php echo $row['id']?>" onclick="return confirm('Are you sure you want to delete the journal article?')">Delete</a></td>
                
 
                </tr>
@@ -90,4 +61,6 @@ if (isset($_POST['search'])){
 
 }
    ?>
+
+<script src="../js/main.js"></script>
